09-genai-customer-support/requirements.txt
# Core AI/ML packages
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.20

# Vector Database
chromadb>=0.4.22

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.31.0

# Document Processing
pypdf2>=3.0.1
python-docx>=1.1.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.3

# Data Handling
pandas>=2.2.0
numpy>=1.26.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.0
typing-extensions>=4.9.0
requests>=2.31.0

# Development Tools
jupyter>=1.0.0
pytest>=8.0.0
pinecone>=7.0.0
langchain-pinecone>=0.0.1
